requirements.txt
tensorflow-gpu==2.1.0
matplotlib==3.1.2
flake8==3.7.9
spacy==2.2.4
easydict==1.9
pyyaml==5.3.1
-e .
wget==3.2
tqdm==4.45.0